ajuste na exibição de imagens do portfolio

+++ admin/portfolios.php
@@ -11,18 +11,20 @@ $dadosEstilo = $portfolio->estilo->exibir();
 <section class="container">
     <h1 class="text-center">Portfolio</h1>
     <p><a class="btn btn-primary" href="portfolio-inserir.php">Inserir portfolio</a></p>
-<?php foreach($dadosPortfolio as $umPortfolio){?>    
+<?php foreach($dadosPortfolio as $umPortfolio){   
+    if($umPortfolio['usuario_id'] === $_SESSION['id']) {?>
         <div id="imagemPortfolio">
             <img src="../assets/images/portfolio/<?=$umPortfolio['imagem']?>" alt="Foto do portfolio">
         </div>
-
+        
         <p><?=$umPortfolio['descricao']?></p> 
-
+        
         <p><?=$dadosEstilo[$umPortfolio['estilo_id'] - 1]['nome']?></p>
-
+        
         <a class="btn btn-warning" href="portfolio-atualizar.php?id=<?=$umPortfolio['id']?>">Atualizar</a>
         <a class="btn btn-danger" href="portfolio-excluir.php?id=<?=$umPortfolio['id']?>">Excluir</a>
-<?php } ?>
+    <?php } 
+} ?>
         
     </section>    
 <?php
